+++ src/main/java/org/zlab/upfuzz/fuzzingengine/Server/Corpus.java
@@ -3,12 +3,10 @@ package org.zlab.upfuzz.fuzzingengine.Server;
 import java.io.File;
 import java.nio.file.Path;
 import java.util.LinkedList;
-import java.util.PriorityQueue;
 import java.util.Queue;
 
 import org.zlab.upfuzz.CommandSequence;
-import org.zlab.upfuzz.fuzzingengine.TestPacket;
-import org.zlab.upfuzz.fuzzingengine.TestPacket.TestPacketComparator;
+import org.zlab.upfuzz.fuzzingengine.Packet.TestPacket;
 import org.zlab.upfuzz.utils.Pair;
 import org.zlab.upfuzz.utils.Utilities;
 
@@ -17,7 +15,7 @@ public class Corpus {
     // PriorityQueue queue = new PriorityQueue<TestPacket>(
     // new TestPacketComparator());
 
-    Queue<Entry> queue = new LinkedList();
+    Queue<CorpusEntry> queue = new LinkedList();
 
     public boolean initCorpus(Path initSeedDirPath) {
         File initSeedDir = initSeedDirPath.toFile();
@@ -29,7 +27,7 @@ public class Corpus {
                 if (commandSequencePair != null) {
                     // Fuzzer.saveSeed(commandSequencePair.left,
                     // commandSequencePair.right);
-                    queue.add(new Entry(commandSequencePair.left,
+                    queue.add(new CorpusEntry(commandSequencePair.left,
                             commandSequencePair.right));
                 }
             }