+++ src/main/java/org/zlab/upfuzz/fuzzingengine/Server/CorpusEntry.java
@@ -2,14 +2,14 @@ package org.zlab.upfuzz.fuzzingengine.Server;
 
 import org.zlab.upfuzz.CommandSequence;
 
-public class Entry {
+public class CorpusEntry {
     // Write commands
     public CommandSequence originalCommandSequence;
     public CommandSequence upgradedCommandSequence;
     // Read Commands
     public CommandSequence validationCommandSequnece;
 
-    public Entry(CommandSequence originalCommandSequence,
+    public CorpusEntry(CommandSequence originalCommandSequence,
             CommandSequence validationCommandSequnece) {
         this.originalCommandSequence = originalCommandSequence;
         this.validationCommandSequnece = validationCommandSequnece;