+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change log
 
 
+## server-client
++ `2022-06-09` *f8985e3* use corpus entry and stacked test packet
+
 ## hdfs
 + `2022-06-08` *e3076b7* fix merge error: 1. cassandra cqlsh daemon socket initialization 2. cassandra executor startup problem
 + `2022-06-08` *3389543* setup log4j